island: Remove circular dep btw ConfigService and PostBreachFilesService
