+++ monkey/monkey_island/cc/services/post_breach_files.py
@@ -2,16 +2,13 @@ import logging
 import os
 from pathlib import Path
 
-# TODO: Remove circular dependency between ConfigService and PostBreachFilesService.
-import monkey_island.cc.services.config
-
 __author__ = "VakarisZ"
 
 logger = logging.getLogger(__name__)
 
 # Where to find file names in config
-PBA_WINDOWS_FILENAME_PATH = ["monkey", "post_breach", "PBA_windows_filename"]
 PBA_LINUX_FILENAME_PATH = ["monkey", "post_breach", "PBA_linux_filename"]
+PBA_WINDOWS_FILENAME_PATH = ["monkey", "post_breach", "PBA_windows_filename"]
 
 
 class PostBreachFilesService:
@@ -45,19 +42,3 @@ class PostBreachFilesService:
         return os.path.join(
             PostBreachFilesService.DATA_DIR, PostBreachFilesService.CUSTOM_PBA_DIRNAME
         )
-
-    @staticmethod
-    def set_config_PBA_files(config_json):
-        """
-        Sets PBA file info in config_json to current config's PBA file info values.
-        :param config_json: config_json that will be modified
-        """
-        if monkey_island.cc.services.config.ConfigService.get_config():
-            linux_filename = monkey_island.cc.services.config.ConfigService.get_config_value(
-                PBA_LINUX_FILENAME_PATH
-            )
-            windows_filename = monkey_island.cc.services.config.ConfigService.get_config_value(
-                PBA_WINDOWS_FILENAME_PATH
-            )
-            config_json["monkey"]["post_breach"]["PBA_linux_filename"] = linux_filename
-            config_json["monkey"]["post_breach"]["PBA_windows_filename"] = windows_filename
